+++ BE-AdminPanel/database/migrations/2021_08_01_000032_create_user_refunds_table.php
@@ -19,7 +19,7 @@ class CreateUserRefundsTable extends Migration
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
 
-            $table->double('amount');
+            $table->decimal('amount', 15, 2);
             $table->date('refund_date');
 
             //reason
